src/requirements.txt
bcrypt==3.2.0
cffi==1.14.5
mysql-connector-python==8.0.25
protobuf==3.17.3
pycparser==2.20
six==1.16.0
